deleted comments, added descritpions, moved classes in the right place, rename classes
