+++ ISAAR.MSolve.Analyzers/ISAAR.MSolve.Analyzers.csproj
@@ -18,4 +18,7 @@
     <ProjectReference Include="..\ISAAR.MSolve.Numerical\ISAAR.MSolve.Numerical.csproj" />
     <ProjectReference Include="..\ISAAR.MSolve.Solvers\ISAAR.MSolve.Solvers.csproj" />
   </ItemGroup>
+  <ItemGroup>
+    <Folder Include="SupportiveClasses\" />
+  </ItemGroup>
 </Project>
